Lease_tracking_and_reimbursement_system
System written for someone who owns a home and a market. It is a system that warns tenants for their homes when they follow the rental status and the lease payment is delayed. A system where customers can track the situation on the market. Developed using C# Windows Form Application and Microsoft Access.